Fix warnings for Xcode 7 beta 4

+++ Partisans/Classes/DecisionMenuItems.m
@@ -350,7 +350,7 @@
 }
 
 
-- (float)menuViewController:(JSKMenuViewController *)menuViewController heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (double)menuViewController:(JSKMenuViewController *)menuViewController heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     if (indexPath.section == DecisionMenuSectionVotes)
     {
